Creates test_tensor_creation_ops test suite
